refactor(error,index): widen FdiFileCountShort.actual + remove encoded_entry bool
Closes #136. Two H4/H5 micro-fixes from the Phase 1 type-design
audit. Both are pure type-shape improvements with no behavior
change in any reachable execution path.

**H4 — silent truncation**
`EncodedFault::FdiFileCountShort.actual` was `u32`, populated via
`entries.len() as u32`. The check that surfaces the variant is
`entries.len() != file_count as usize`, where `file_count: u32`.
In adversarial-future terms (no current invariant ever exceeds
`u32::MAX` here), `entries.len() > u32::MAX` would truncate to a
value that could equal `file_count`, producing a self-
contradictory diagnostic ("actual=N but file_count claims N").

Widened `actual` to `u64`. The `entries.len() as u64` cast at the
construction site is a lossless widening on every supported
target. Two test pattern-matches at `index/mod.rs:3045` and
:3692 keep their numeric literals (`1`, `0`) — they auto-infer
to the new field type.

**H5 — boolean blindness**
`encoded_entry_in_data_record_size(compressed: bool, block_count:
usize)` carried a single naked `bool` at a load-bearing internal
seam — the same anti-pattern that `EntryFlags` (container/mod.rs:
49) was extracted to defeat, in miniature. The production call
site `encoded_entry_in_data_record_size(compression_method !=
CompressionMethod::None, block_count as usize)` is exactly the
shape that breaks under a future second-bool addition (e.g.
`with_aes_alignment`).

Changed signature to `(method: &CompressionMethod, block_count:
usize)` and compute the compressed branch internally via
`!matches!(method, CompressionMethod::None)`. The bool is gone
entirely — a future bool addition next to `method` can't be
swapped because the types differ.

Dropped the `const fn` qualifier (the function is not called in
any const context, and the `matches!` pattern is non-const for
non-Copy enum variants like `CompressionMethod::Unknown(_)`).

Touched: 1 production call site (entry_header.rs:362) + 8 test
sites (mod.rs:1411 proptest comparator + 3 helper-style tests +
4 pin tests). Construction shape change is mechanical:
`(true, N)` → `(&CompressionMethod::Zlib, N)`, `(false, N)` →
`(&CompressionMethod::None, N)`.

Validation:
- cargo test --workspace --all-features: 489 tests pass
- cargo clippy --workspace --all-targets --all-features -D warnings: clean
- cargo minimal-versions check --direct --all-features: clean
- cargo fmt --all --check: clean (after re-flowing 4 long lines)

+++ crates/paksmith-core/src/container/pak/index/path_hash.rs
@@ -722,7 +722,11 @@ impl PakIndex {
                 fault: IndexParseFault::Encoded {
                     kind: EncodedFault::FdiFileCountShort {
                         file_count,
-                        actual: entries.len() as u32,
+                        // entries.len() is usize; on every supported
+                        // Rust target usize <= u64, so `as u64` is a
+                        // lossless widening cast (issue #136 retired
+                        // the `as u32` that could truncate).
+                        actual: entries.len() as u64,
                     },
                 },
             });
